Minor changes and improvements
Add the "scripts" folder to store the box64 and box86 install scripts there, although I plan to use it to add more scripts.

I added the variable "MAX_PLAYERS" in all the eggs, that will allow to adjust the number of maximum people that can enter a server (adjustable only by administrators).

I removed a couple of unnecessary lines in the MTA install script (in the egg).

And lastly I changed the URL where the docker image will download the "install-box64.sh" and "install-box86.sh" files.