Big update
I created some scripts that will start the servers. These scripts, in addition to starting the server, will check the architecture of the machine to be able to know whether to execute the start command for ARM or AMD, in addition to that, the IP address of the server can be established, as well as the port through flags, as well A limit may be placed on the maximum players per server. The scripts will also check if the users are using the established data and limits and if not, they will be notified that the data and limits have been changed to the established ones.

I edited the eggs and yolks to these files, also to save space I made the folder where all that garbage for installation and construction was deleted when installing box86.